create product initial
