+++ components/admin/product/CreateProduct.tsx
@@ -1,0 +1,6 @@
+import { Title } from "@mantine/core";
+
+const CreateProduct = () => {
+  return <Title>HEJ!</Title>;
+};
+export default CreateProduct;
