+++ pages/api/admin/subproduct/index.ts
@@ -75,29 +75,14 @@ export default async function handler(
       try {
         // Fix validation for already existing sub product except the one you update
 
-        // Check i Main product exists
-        /*       const mainProduct: MainProductDocument | null =
-          await MainProduct.findOne({
-            _id: req.body.mainProduct,
-          });
-
-        if (!mainProduct) {
-          return res.status(404).send({
-            success: false,
-            data: "Main product could not be found",
-          });
-        } */
-
         const todayDate = new Date()
           .toISOString()
           .slice(0, 16)
           .replace("T", " ");
-        //console.log(req.body);
+
         const updateSubProduct: SubProductDocument = req.body;
-        //updateSubProduct.setSlug!(req.body.title);
         updateSubProduct.lastUpdated = todayDate;
 
-        console.log(updateSubProduct);
         const subProduct = await SubProduct.findOneAndUpdate(
           { _id: req.body._id },
           updateSubProduct,
@@ -107,7 +92,6 @@ export default async function handler(
           }
         );
 
-        console.log(subProduct);
         if (!subProduct) {
           return res
             .status(400)
